Use float zoom to calculate visible bounds
Use the original float zoom value to calculate the visible bounds in NonHierarchicalViewBasedAlgorithm so that the bounds don't jump between sizes for only discrete integer zoom levels. This avoids the visible bounds being unnecessarily large for non-discrete zoom levels.

+++ library/src/com/google/maps/android/clustering/algo/Algorithm.java
@@ -34,7 +34,7 @@ public interface Algorithm<T extends ClusterItem> {
 
     void removeItem(T item);
 
-    Set<? extends Cluster<T>> getClusters(double zoom);
+    Set<? extends Cluster<T>> getClusters(float zoom);
 
     Collection<T> getItems();
